`Renew`: Support DocumentDB and Shared Lock Bugfix
Through testing I have discovered that AWS Document does not seem to like the BSON produced when directly using `time.Time` in `FindOneAndUpdate`. Switching to using `primitive.NewDateTimeFromTime` works with both MongoDB and DocumentDB.

Updated the selector for shared locks in `Renew` to use `$elemMatch` instead of dot notation. Per the [Mongodb documenation](https://www.mongodb.com/docs/manual/tutorial/query-array-of-documents/#specify-multiple-conditions-for-array-of-documents) the dot notation with respect to array elements only ensures that some combination of array elements in the document meet the query requirements. `$elemMatch` requires that at least one array element meets _all_ of the query requirements, which is what we want to have happen in `Renew`. The updates to the `TestRenew` unit test demonstrate this. If run using dot notation the lock `cccc` is renewed instead of `aaaa` which is not the intended behavior.